Learn Interactive 2016
The files in this folder, (when running from cmd use
student_homepage.py) have most of the functionality needed for the
application. The only things we need to do are test feedback (Kieren),
Lecturer homepage (Ben), Login/Registration system (Matteo) and making
it look nicer (everyone). Also we plan to run through a demo during a
free period on Thursday so if everyone could upload their final parts by
Wednesday night that would be good :)